% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{cvm.testg}
\alias{cvm.testg}
\title{Goodness of Fit Tests}
\usage{
ad.testg(x,cdf)
cvm.testg(x,cdf)
}
\arguments{
\item{x}{a numeric vector of data values, the number of which must be greater than 7. Missing values are allowed.}

\item{cdf}{cumulative distribution function.}
}
\value{
A list with class "htest" containing the following components:

\code{statistic}	the value of the test statistic.

\code{p.value}	the p-value for the test.

\code{method}	a character string indicating what type of likelihood ratio test was performed.

\code{data.name} a character string giving the name(s) of the data
}
\description{
Performs the Anderson-Darlin and Cramer-von Mises tests
}
\examples{
x<- rRBS(1000)
fit <- gamlss(x~1,family=RBS(),method=CG())
mu<- fit$mu.coefficients ; mu
sigma <- fit$sigma.coefficients ; sigma
cdf <- function(x) pRBS(x,mu,sigma)
ad.testg(x,cdf)
cvm.testg(x,cdf)
}
\references{
Stephens, M.A. (1986): Tests based on EDF statistics. In: D'Agostino, R.B. and Stephens, M.A., eds.: Goodness-of-Fit Techniques. Marcel Dekker, New York.

Thode Jr., H.C. (2002): Testing for Normality. Marcel Dekker, New York.

Chen, G., Balakrishnan, N. (1995) A general purpose approximate goodness-of-fit test,  \emph{J. Quality Technol.},  v. 27, 154-161.

Barros, M., Leiva, V., Ospina, R. ; Tsuyuguchi, A.B. (2014) Goodness-of-fit tests based on the Birnbaum-Saunders distribution for censored reliability data analysis. \emph{IEEE Transactions on Reliability}, v. 63, p. 543/554.
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
