% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{diag.bs}
\alias{diag.bs}
\title{Diagnostic Analysis - Local Influnce}
\usage{
diag.bs(model)
}
\arguments{
\item{model}{object of class \code{gamlss} holding the fitted model.}
}
\value{
Local influence measures.
}
\description{
Diagnostics for the RBS model
}
\examples{
library(faraway)
data(cpd)
attach(cpd)
fit = gamlss(actual ~ 0+projected, family=RBS(mu.link="identity"),method=CG())
summary(fit)
set.seed(2015)
envelope(fit,precision="fixed")
Cib <- diag.bs(fit)$Ci.beta
plot(Cib,ylim=c(0,1),pch=19,ylab=expression(C[i](beta)),xlab="Index")
abline(h=2*mean(Cib),lty=2)
Cia <- diag.bs(fit)$Ci.alpha
plot(Cia,ylim=c(0,1),pch=19,ylab=expression(C[i](beta)),xlab="Index")
abline(h=2*mean(Cia),lty=2)
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
\references{
Leiva, V., Santos-Neto, M., Cysneiros, F.J.A, Barros, M. (2014)  Birnbaum-Saunders statistical modelling: a new approach. \emph{Statistical Modelling}, v. 14, p. 21-48, 2014.
}

