% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{diag.RBS}
\alias{diag.RBS}
\title{Diagnostic Analysis - Local Influnce}
\usage{
diag.RBS(
  model,
  mu.link = "identity",
  sigma.link = "identity",
  scheme = "case.weight",
  lx = NULL,
  lz = NULL
)
}
\arguments{
\item{model}{Object of class \code{gamlss} holding the fitted model.}

\item{mu.link}{Defines the mu.link, with "identity" link as the default for the mu parameter.}

\item{sigma.link}{Defines the sigma.link, with "identity" link as the default for the sigma parameter}

\item{scheme}{Default is "case.weight". But, can be "response", "location" or "precision".}

\item{lx}{Used in the scheme 'location'.}

\item{lz}{Used in the scheme 'precision'.}
}
\value{
Local influence measures.
}
\description{
Diagnostics for the RBS model
}
\examples{
data(cpd,package='faraway')
attach(cpd)
fit = gamlss::gamlss(actual ~ 0+projected,
family=RBS(mu.link="identity",sigma.link="identity"),method=CG())
Cib <- diag.RBS(fit)$Ci.beta
plot(Cib,ylim=c(0,1),pch=19,ylab=expression(C[i](beta)),xlab="Index")
abline(h=2*mean(Cib),lty=2)
Cia <- diag.RBS(fit)$Ci.alpha
plot(Cia,ylim=c(0,1),pch=19,ylab=expression(C[i](beta)),xlab="Index")
abline(h=2*mean(Cia),lty=2)


}
\references{
Leiva, V., Santos-Neto, M., Cysneiros, F.J.A, Barros, M. (2014)  Birnbaum-Saunders statistical modelling: a new approach. \emph{Statistical Modelling}, v. 14, p. 21-48, 2014.
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
