% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelopebs.R
\name{envelope}
\alias{envelope}
\alias{envelope.ZARBS}
\alias{envelope.ZAGA}
\title{Envelopes}
\usage{
envelope(
  model,
  k = 100,
  res = "deviance",
  precision = "fixed",
  dist = RBS(mu.link = "identity", sigma.link = "identity"),
  color = "grey50",
  xlabel = "Theorical Quantile",
  ylabel = "Empirical Quantile",
  font = "serif"
)

envelope.ZARBS(
  model,
  k = 100,
  precision = "fixed",
  color = "grey50",
  xlabel = "Theorical Quantile",
  ylabel = "Empirical Quantile",
  font = "serif"
)

envelope.ZAGA(
  model,
  k = 100,
  precision = "fixed",
  color = "grey50",
  xlabel = "Theorical Quantile",
  ylabel = "Empirical Quantile",
  font = "serif"
)
}
\arguments{
\item{model}{object of class \code{gamlss} holding the fitted model.}

\item{k}{number of replications for envelope construction. Default is 19.}

\item{res}{type of residuals to be extracted. Default is deviance.}

\item{precision}{If \code{precision = "fixed"} a model with precision fixed is used;
else a model with precision variable is used.}

\item{dist}{The function RBS() defines the RBS distribution.}

\item{color}{the color of the envelope.}

\item{xlabel}{a label for the x axis.}

\item{ylabel}{a label for the y axis.}

\item{font}{the font used in x and y axis.}
}
\value{
A simulated envelope of the class RBS, ZARBS and ZAGA.
}
\description{
A normal plot with simulated envelope of the residual is produced.
}
\references{
Atkinson, A. C. (1985) Plots, transformations and regression : an introduction to graphical methods of diagnostic regression analysis. Oxford Science Publications, Oxford.

Leiva, V., Santos-Neto, M., Cysneiros, F.J.A, Barros, M. (2014)  Birnbaum-Saunders statistical modelling: a new approach. \emph{Statistical Modelling}, v. 14, p. 21-48, 2014.

Santos-Neto, M., Cysneiros, F.J.A., Leiva, V., Barros, M. (2016) Reparameterized Birnbaum-Saunders
regression models with varying precision. \emph{Electronic Journal of Statistics}, 10, 2825--2855. doi: \email{10.1214/16-EJS1187}.
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
