% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbs.R
\name{est.rbs}
\alias{est.rbs}
\title{Reparameterized Birnbaum-Saunders (RBS) distribution for fitting a GAMLSS}
\usage{
est.rbs(x,xi=0.95)
}
\arguments{
\item{x}{vector of data.}

\item{xi}{confidence level. Default is 0.95.}
}
\value{
Returns the estimates and confidence intervals of mu and sigma.
}
\description{
The fuction \code{RBS()} defines the BS distribution, a two paramenter
distribution, for a gamlss.family object to be used in GAMLSS fitting using using the
function \code{gamlss()}, with mean equal to the parameter \code{mu} and \code{sigma}
equal the precision parameter. The functions \code{dRBS}, \code{pRBS}, \code{qRBS} and
\code{rBS} define the density, distribution function, quantile function and random
genetation for the \code{RBS} parameterization of the RBS distribution.
}
\details{
The parametrization of the normal distribution given in the function RBS() is

\deqn{f_{Y}(y;\mu,\sigma)=\frac{\exp\left(\sigma/2\right)\sqrt{\sigma+1}}{4\sqrt{\pi\mu}\,y^{3/2}}
\left[y+\frac{\sigma \mu}{\sigma+1}\right] \exp\left(-\frac{\sigma}{4}
                                                    \left[\frac{y\{\sigma+1\}}{\sigma\mu}+\frac{\sigma\mu}{y\{\sigma+1\}}\right]\right) y>0.}
}
\examples{
data(psi31)
est.rbs(psi31)

x<- rRBS(100)
est.rbs(x)
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}

