% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbs.R, R/zabs.R
\name{Ims}
\alias{Ims}
\title{Reparameterized Birnbaum-Saunders (RBS) distribution for fitting a GAMLSS}
\usage{
RBS(mu.link = "identity", sigma.link = "identity")
dRBS(x, mu = 1, sigma = 1, log = FALSE)
pRBS(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
qRBS(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
rRBS(n, mu = 1, sigma = 1)
plotRBS(mu = .5, sigma = 1, from = 0, to = 0.999, n = 101, ...)
meanRBS(obj)

ZARBS(mu.link = "identity", sigma.link = "identity")
dZARBS(x, mu = 1, sigma = 1, nu=0.1, log = FALSE)
pZARBS(q, mu = 1, sigma = 1,  nu=0.1, lower.tail = TRUE, log.p = FALSE)
qZARBS(p, mu = 1, sigma = 1,  nu=0.1, lower.tail = TRUE, log.p = FALSE)
rZARBS(n, mu = 1, sigma = 1)
plotZARBS(mu = .5, sigma = 1,  nu=0.1, from = 0, to = 0.999, n = 101, ...)
meanZARBS(obj)
}
\arguments{
\item{mu}{vector of scale parameter values}

\item{sigma}{vector of shape parameter values}

\item{mu.link}{object for which the extraction of model residuals is meaningful.}

\item{sigma.link}{type of residual to be used.}

\item{x, q}{vector of quantiles}

\item{log, }{log.p logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{from}{where to start plotting the distribution from}

\item{to}{up to where to plot the distribution}

\item{obj}{a fitted ZARBS object}

\item{...}{other graphical parameters for plotting}
}
\value{
returns a \code{gamlss.family} object which can be used to fit a normal distribution in the \code{gamlss()} function.

returns a \code{gamlss.family} object which can be used to fit a normal distribution in the \code{gamlss()} function.
}
\description{
The fuction \code{RBS()} defines the BS distribution, a two paramenter
distribution, for a gamlss.family object to be used in GAMLSS fitting using using the
function \code{gamlss()}, with mean equal to the parameter \code{mu} and \code{sigma}
equal the precision parameter. The functions \code{dRBS}, \code{pRBS}, \code{qRBS} and
\code{rBS} define the density, distribution function, quantile function and random
genetation for the \code{RBS} parameterization of the RBS distribution.

The fuction \code{ZARBS()} defines the ZARBS distribution, a two paramenter
distribution, for a gamlss.family object to be used in GAMLSS fitting using using the
function \code{gamlss()}. The zero adjusted Birnbaum-Saunders distribution is similar
to the Birnbaum-Saunders distribution but allows zeros as y values. The extra parameter
models the probabilities at zero. The functions dZARBS, pZARBS, qZARBS and rZARBS define
the density, distribution function, quantile function and random generation for
the ZARBS. plotZARBS can be used to plot the distribution. meanZARBS calculates the expected
value of the response for a fitted model.
}
\details{
The parametrization of the normal distribution given in the function RBS() is

\deqn{f_{Y}(y;\mu,\sigma)=\frac{\exp\left(\sigma/2\right)\sqrt{\sigma+1}}{4\sqrt{\pi\mu}\,y^{3/2}}
\left[y+\frac{\sigma \mu}{\sigma+1}\right] \exp\left(-\frac{\sigma}{4}
                                                    \left[\frac{y\{\sigma+1\}}{\sigma\mu}+\frac{\sigma\mu}{y\{\sigma+1\}}\right]\right) y>0.}

The parametrization of the zero adjusted reparameterized Birnbaum-Saunders distribution given in the function ZARBS() is

\deqn{ f_{Y}(y;\mu,\delta,p) =\frac{[1-p]\sqrt{\delta+1}}{4\,y^{3/2}\,\sqrt{\pi\mu}}\left[y+\frac{\delta\mu}{\delta+1} \right]\exp\left(-\frac{\delta}{4}\left[\frac{y[\delta+1]}{\delta\mu}+\frac{\delta\mu}{y[\delta+1]}-2\right]\right) I_{(0, \infty)}(y)+  pI_{\{0\}}(y).}
}
\note{
For the function RBS(), mu is the mean and sigma is the precision parameter of the Birnbaum-Saunders distribution.

For the function ZARBS(), mu is the mean and sigma is the precision parameter and nu is the proportion of zeros of the ZARBS distribution.
}
\examples{
plotRBS()
dat <- rRBS(1000); hist(dat)
fit <- gamlss(dat~1,family=RBS(),method=CG())
meanRBS(fit)

##install.packages(faraway)
library(faraway)
data(cpd)
attach(cpd)
model0 = gamlss(actual ~ projected, family=RBS(mu.link="identity"),method=CG())
summary(model0)
model = gamlss(actual ~ 0+projected, family=RBS(mu.link="identity"),method=CG())
summary(model)
plotZABS()
dat <- rZARBS(1000); hist(dat)
fit <- gamlss(dat~1,family=ZARBS(),method=CG())
meanZABS(fit)

data(papatoes);
fit = gamlss(I(Demand/10000)~1,sigma.formula=~1, nu.formula=~1, family=ZARBS(mu.link="identity",sigma.link = "identity",nu.link = "identity"),method=CG(),data=papatoes)
summary(fit)

data(oil)
fit1 = gamlss(Demand~1,sigma.formula=~1, nu.formula=~1, family=ZARBS(mu.link="identity",sigma.link = "identity",nu.link = "identity"),method=CG(),data=oil)
summary(fit1)

}
\references{
Santos-Neto, M., Cysneiros, F.J.A, Leiva, V., Barros, M. (2012) On new parameterizations of the Birnbaum-Saunders distribution. \emph{PAK J STAT}, v. 28, p. 1-26, 2012.

Santos-Neto, M., Cysneiros, F.J.A, Leiva, V., Barros, M. (2014) On a reparameterized Birnbaum-Saunders distribution and its moments, estimation and application. \emph{Revstat Statistical Journal}, v. 12, p. 247-272, 2014.

Leiva, V., Santos-Neto, M., Cysneiros, F.J.A, Barros, M. (2014)  Birnbaum-Saunders statistical modelling: a new approach. \emph{Statistical Modelling}, v. 14, p. 21-48, 2014.

Leiva, V., Santos-Neto, M., Cysneiros, F.J.A., Barros, M. (2016) A methodology for stochastic inventory models based on a zero-adjusted Birnbaum-Saunders distribution.
\emph{Applied Stochastic Models in Business and Industry.}, 32(1), 74--89. doi:\email{10.1002/asmb.2124}.

Santos-Neto, M., Cysneiros, F.J.A., Leiva, V., Barros, M. (2016) Reparameterized Birnbaum-Saunders
regression models with varying precision. \emph{Electronic Journal of Statistics}, 10, 2825--2855. doi: \email{10.1214/16-EJS1187}.
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}

Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
