% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zabs.R
\name{meanZARBS}
\alias{meanZARBS}
\title{Zero Adjusted Reparameterized Birnbaum-Saunders (ZARBS) distribution for fitting a GAMLSS}
\usage{
ZARBS(mu.link = "identity", sigma.link = "identity")
dZARBS(x, mu = 1, sigma = 1, nu=0.1, log = FALSE)
pZARBS(q, mu = 1, sigma = 1,  nu=0.1, lower.tail = TRUE, log.p = FALSE)
qZARBS(p, mu = 1, sigma = 1,  nu=0.1, lower.tail = TRUE, log.p = FALSE)
rZARBS(n, mu = 1, sigma = 1)
plotZARBS(mu = .5, sigma = 1,  nu=0.1, from = 0, to = 0.999, n = 101, ...)
meanZARBS(obj)
}
\arguments{
\item{obj}{a fitted ZARBS object}

\item{mu.link}{object for which the extraction of model residuals is meaningful.}

\item{sigma.link}{type of residual to be used.}

\item{x, q}{vector of quantiles}

\item{mu}{vector of scale parameter values}

\item{sigma}{vector of shape parameter values}

\item{log, }{log.p logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{from}{where to start plotting the distribution from}

\item{to}{up to where to plot the distribution}

\item{...}{other graphical parameters for plotting}
}
\value{
returns a \code{gamlss.family} object which can be used to fit a normal distribution in the \code{gamlss()} function.
}
\description{
The fuction \code{ZARBS()} defines the ZARBS distribution, a two paramenter
distribution, for a gamlss.family object to be used in GAMLSS fitting using using the
function \code{gamlss()}. The zero adjusted Birnbaum-Saunders distribution is similar
to the Birnbaum-Saunders distribution but allows zeros as y values. The extra parameter
models the probabilities at zero. The functions dZARBS, pZARBS, qZARBS and rZARBS define
the density, distribution function, quantile function and random generation for
the ZARBS. plotZARBS can be used to plot the distribution. meanZARBS calculates the expected
value of the response for a fitted model.
}
\details{
The parametrization of the zero adjusted reparameterized Birnbaum-Saunders distribution given in the function ZARBS() is

\deqn{ f_{Y}(y;\mu,\delta,p) =\frac{[1-p]\sqrt{\delta+1}}{4\,y^{3/2}\,\sqrt{\pi\mu}}\left[y+\frac{\delta\mu}{\delta+1} \right]\exp\left(-\frac{\delta}{4}\left[\frac{y[\delta+1]}{\delta\mu}+\frac{\delta\mu}{y[\delta+1]}-2\right]\right) I_{(0, \infty)}(y)+  pI_{\{0\}}(y).}
}
\note{
For the function ZARBS(), mu is the mean and sigma is the precision parameter and nu is the proportion of zeros of the ZARBS distribution.
}
\examples{
plotZABS()
dat <- rZARBS(1000); hist(dat)
fit <- gamlss(dat~1,family=ZARBS(),method=CG())
meanZABS(fit)

data(papatoes);
fit = gamlss(I(Demand/10000)~1,sigma.formula=~1, nu.formula=~1, family=ZARBS(mu.link="identity",sigma.link = "identity",nu.link = "identity"),method=CG(),data=papatoes)
summary(fit)

data(oil)
fit1 = gamlss(Demand~1,sigma.formula=~1, nu.formula=~1, family=ZARBS(mu.link="identity",sigma.link = "identity",nu.link = "identity"),method=CG(),data=oil)
summary(fit1)

}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
\references{
Leiva, V., Santos-Neto, M., Cysneiros, F.J.A., Barros, M. (2016) A methodology for stochastic inventory models based on a zero-adjusted Birnbaum-Saunders distribution.
\emph{Applied Stochastic Models in Business and Industry.}, 32(1), 74--89. doi:\email{10.1002/asmb.2124}.

Santos-Neto, M., Cysneiros, F.J.A., Leiva, V., Barros, M. (2016) Reparameterized Birnbaum-Saunders
regression models with varying precision. \emph{Electronic Journal of Statistics}, 10, 2825--2855. doi: \email{10.1214/16-EJS1187}.
}

