% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs2.R
\docType{data}
\name{psi26}
\alias{psi26}
\title{Lifetimes of aluminum specimens exposed to a maximum stress of 26,000 psi.}
\format{
A vector containing 102 observations.
}
\usage{
data(psi26)
}
\description{
Several data sets related to the BS distribution are available in the RBS package, which have been taken from the literature on this topic.
}
\details{
psi21, psi26 and psi31 were taken from Birnbaum and Saunders (1969),
who reported fatigue life data correspond to the cycles (\eqn{\times 10^{-3}}) of aluminum
specimens of type 6061-T6. These specimens were cut in a parallel angle to the direction of
rotation and oscillating at 18 cycles per seconds. They were exposed to a pressure with maximum
stress of 21,000, 26,000, and 31,000 psi (pounds per square inch), for \eqn{n = 101, 102,}
and \eqn{101} specimens, respectively.
}
\examples{
## Load data sets
data(psi26)
## Histogram for psi26
hist(psi26)

}
\references{
Birnbaum, Z. W. and Saunders, S. C. (1969). Estimation for a family of life distributions with applications to fatigue. J. Appl. Probab. 6(2): 328-347.
}
\keyword{datasets}
