% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbs.R
\name{RBS}
\alias{RBS}
\alias{dRBS}
\alias{pRBS}
\alias{qRBS}
\alias{rRBS}
\alias{plotRBS}
\alias{meanRBS}
\alias{sigmatil}
\alias{resrbs}
\alias{est.rbs}
\alias{Ims}
\title{Reparameterized Birnbaum-Saunders (RBS) distribution for fitting a GAMLSS}
\usage{
RBS(mu.link = "log", sigma.link = "log")

sigmatil(y)

Ims(mu, sigma)

resrbs(y, mu, sigma)

dRBS(x, mu = 1, sigma = 1, log = FALSE)

pRBS(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qRBS(p, mu = 0.5, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rRBS(n, mu = 1, sigma = 1)

plotRBS(
  mu = 0.5,
  sigma = 1,
  from = 0,
  to = 0.999,
  n = 101,
  title = "title",
  ...
)

meanRBS(obj)

est.rbs(x, xi = 0.95)
}
\arguments{
\item{mu.link}{object for which the extraction of model residuals is meaningful.}

\item{sigma.link}{type of residual to be used.}

\item{mu}{vector of scale parameter values}

\item{sigma}{vector of shape parameter values}

\item{x, y, q}{vector of quantiles}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{from}{where to start plotting the distribution from.}

\item{to}{up to where to plot the distribution.}

\item{title}{title of the plot.}

\item{...}{other graphical parameters for plotting.}

\item{obj}{a fitted RBS object,}

\item{xi}{the confidence level.}
}
\value{
returns a \code{gamlss.family} object which can be used to fit a normal distribution in the \code{gamlss()} function.
}
\description{
The fuction \code{RBS()} defines the BS distribution, a two paramenter
distribution, for a gamlss.family object to be used in GAMLSS fitting using using the
function \code{gamlss()}, with mean equal to the parameter \code{mu} and \code{sigma}
equal the precision parameter. The functions \code{dRBS}, \code{pRBS}, \code{qRBS} and
\code{rBS} define the density, distribution function, quantile function and random
genetation for the \code{RBS} parameterization of the RBS distribution.
}
\details{
The parametrization of the normal distribution given in the function RBS() is

\deqn{f_{Y}(y;\mu,\sigma)=\frac{\exp\left(\sigma/2\right)\sqrt{\sigma+1}}{4\sqrt{\pi\mu}\,y^{3/2}}
\left[y+\frac{\sigma \mu}{\sigma+1}\right] \exp\left(-\frac{\sigma}{4}
                                                    \left[\frac{y\{\sigma+1\}}{\sigma\mu}+\frac{\sigma\mu}{y\{\sigma+1\}}\right]\right) y>0.}
}
\note{
For the function RBS(), mu is the mean and sigma is the precision parameter of the Birnbaum-Saunders distribution.
}
\examples{
data(cpd,package='faraway')
attach(cpd)
model0 = gamlss::gamlss(actual ~ projected, family=RBS(mu.link="identity"),method=CG())
summary(model0)
model = gamlss::gamlss(actual ~ 0+projected, family=RBS(mu.link="identity"),method=CG())
summary(model)

}
\references{
Santos-Neto, M., Cysneiros, F.J.A, Leiva, V., Barros, M. (2012) On new parameterizations of the Birnbaum-Saunders distribution. \emph{PAK J STAT}, v. 28, p. 1-26, 2012.

Santos-Neto, M., Cysneiros, F.J.A, Leiva, V., Barros, M. (2014) On a reparameterized Birnbaum-Saunders distribution and its moments, estimation and application. \emph{Revstat Statistical Journal}, v. 12, p. 247-272, 2014.

Leiva, V., Santos-Neto, M., Cysneiros, F.J.A, Barros, M. (2014)  Birnbaum-Saunders statistical modelling: a new approach. \emph{Statistical Modelling}, v. 14, p. 21-48, 2014.
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
