% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{residuals}
\alias{residuals}
\title{Residuals}
\usage{
residuals(model,residual = "deviance")
}
\arguments{
\item{model}{an object for which the extraction of model residuals is meaningful.}

\item{residual}{type of residual to be used.}
}
\value{
Residuals extracted from the object object.
}
\description{
\code{residuals} is a function which extracts model residuals from objects returned by modeling functions.
}
\examples{
##
library(alr3)
data(landrent)
attach(landrent)
resp <- I(Y/X1)
y1 <-  split(resp, X4)$"1"
x21 <-  split(X2, X4)$"1"

##Fixed Precision
fit0 <- gamlss(y1 ~ x21, family=RBS(mu.link="identity"),method=CG()  )
plot(fitted(fit0),residuals(fit0),xlab="fitted values",ylab="Deviance")
##Varying Precision
fit1 <- gamlss(y1 ~ x21,sigma.formula = y1 ~x21, family=RBS(mu.link="identity",sigma.link="sqrt"),method=CG()  )
plot(fitted(fit1),residuals(fit1),xlab="fitted values",ylab="Deviance")

}
\author{
Manoel Santos-Neto \url{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \url{cysneiros@de.ufpe.br}, Victor Leiva \url{victorleivasanchez@gmail.com} and Michelli Barros \url{michelli.karinne@gmail.com}
}
\references{
Leiva, V., Santos-Neto, M., Cysneiros, F.J.A, Barros, M. (2014)  Birnbaum-Saunders statistical modelling: a new approach. \emph{Statistical Modelling}, v. 14, p. 21-48, 2014.
}

