% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{wald.test}
\alias{wald.test}
\title{Precision test}
\usage{
lr.test(modelh0,modelh1)
grad.test(modelh0,modelh1)
wald.test(modelh0,modelh1)
score.test(modelh0,modelh1)
}
\arguments{
\item{modelh1}{model under alternative hypothesis.}

\item{modelh0}{model under null hypothesis.}
}
\value{
A list with class "htest" containing the following components:

\code{statistic}	the value of the test statistic.

\code{parameter}	the degrees of freedom for the test statistic.

\code{p.value}	the p-value for the test.

\code{method}	a character string indicating what type of likelihood ratio test was performed.

\code{data.name} a character string giving the name(s) of the data
}
\description{
Tests the null hypothesis of precision fixed in RBS models against the alternative of precision variable.
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
#'
}

