% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{diag.ZARBS}
\alias{diag.ZARBS}
\alias{dias.ZARBS}
\title{Diagnostic Analysis - Local Influence}
\usage{
diag.ZARBS(model, links = c("log", "identity", "probit"))
}
\arguments{
\item{model}{object of class \code{gamlss} holding the fitted model.}

\item{links}{link function used.}
}
\description{
Diagnostics for the ZARBS model
}
\references{
Leiva, V., Santos-Neto, M., Cysneiros, F.J.A, Barros, M. (2014)  Birnbaum-Saunders statistical modelling: a new approach. \emph{Statistical Modelling}, v. 14, p. 21-48, 2014.
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
